@@ -4,7 +4,7 @@
   "type": "library",
   "homepage": "https://github.com/jgauthi/component_database",
   "require": {
-    "php": ">=5.6",
+    "php": ">=7.4",
     "ext-mbstring": "*",
     "ext-pdo": "*"
   },
