@@ -1,6 +1,5 @@
 <?php
-use Jgauthi\Component\Database\NavigationPage;
-use Jgauthi\Component\Database\PdoUtils;
+use Jgauthi\Component\Database\{NavigationPage, PdoUtils};
 
 // In this example, the vendor folder is located in "example/"
 require_once __DIR__.'/vendor/autoload.php';
